fix:기본 세팅 변경
